Update patch set 1
Patch Set 1:

Looking at the patch that caused the FTBFS, the issue seems to be caused by this snippet:

        args = {
            'file_path': file_path,
            'section_regex': r'^(\w+)',
            'option_regex': r'^(?:\s+)(\w+(?:\s+\w+)*?)\s+([\w/]*)$'
        }

        self.h_conf.parse_haproxy_conf(file_path)
        mock_generic_ini_style_conf_parser.assert_called_once_with(*args.values())

With Python 3, the order of *args.values() matches the order in which they are described in args, however this is not the case for Python 2, causing the mismatch seen in the log.

If we change the last line to:

        mock_generic_ini_style_conf_parser.assert_called_once_with(
            args['file_path'], args['section_regex'], args['option_regex'])

it works fine on both Python 2 and Python 3.

Patch-set: 1
Reviewer: Gerrit User 113 <113@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Workflow=0